Add intro sections to validator page.
