+++ src/web/src/browser/views/components/ValidatorFileSelectForm.tsx
@@ -1,0 +1,62 @@
+import React from 'react';
+
+import { onFileInputChangeGetFile } from '../../util/file-input';
+import { useActions, useAppState } from '../hooks';
+
+export const ValidatorFileSelectForm = () => {
+  const { sourceRepository, schematron } = useAppState();
+  const actions = useActions();
+  return (
+    <>
+      <div className="tablet:grid-col-4">
+        <div className="usa-hint" id="file-input-specific-hint">
+          Select FedRAMP OSCAL SSP XML file
+        </div>
+        <input
+          id="file-input-specific"
+          className="usa-file-input"
+          type="file"
+          name="file-input-specific"
+          aria-describedby="file-input-specific-hint"
+          accept=".xml"
+          onChange={onFileInputChangeGetFile(fileDetails => {
+            actions.validator.setXmlContents({
+              fileName: fileDetails.name,
+              xmlContents: fileDetails.text,
+            });
+          })}
+          disabled={schematron.validator.current === 'PROCESSING'}
+        />
+      </div>
+      <div className="tablet:grid-col-8">
+        <div className="usa-hint">
+          Or use an example file, brought to you by FedRAMP:
+        </div>
+        <ul className="usa-icon-list margin-top-2">
+          {sourceRepository.sampleSSPs.map((sampleSSP, index) => (
+            <li key={index} className="usa-icon-list__item">
+              <div className="usa-icon-list__icon text-primary">
+                <svg className="usa-icon" aria-hidden="true" role="img">
+                  <use
+                    xlinkHref={actions.getAssetUrl(
+                      'uswds/img/sprite.svg#upload_file',
+                    )}
+                  ></use>
+                </svg>
+              </div>
+              <div className="usa-icon-list__content">
+                <button
+                  className="usa-button usa-button--unstyled"
+                  onClick={() => actions.validator.setXmlUrl(sampleSSP.url)}
+                  disabled={schematron.validator.current === 'PROCESSING'}
+                >
+                  {sampleSSP.displayName}
+                </button>
+              </div>
+            </li>
+          ))}
+        </ul>
+      </div>
+    </>
+  );
+};
